Curry and uncurry unary functions
Curry for a unary function is clearly the identity function.
Action can be considered as special form of Func where return type is
always Void.
Therefore a unary Action is just a special case of a unary Func, hence
Curry for a unary Action is also the identity function.
(And not the function (a -> void) -> (a -> (void -> void)) as previously
implemented).
A similar argument can be made for Uncurry.